services/code-execution-service/controller/requirements.txt
fastapi
kubernetes
pydantic
boto3
pyyaml
